+++ grass-diary/src/components/Feed.jsx
@@ -1,0 +1,75 @@
+import * as stylex from '@stylexjs/stylex';
+import { Link } from 'react-router-dom';
+
+const feed = stylex.create({
+  box: {
+    backgroundColor: '#EAEAEA',
+    borderRadius: '20px',
+    margin: '10px',
+    padding: '20px 30px',
+    width: '360px',
+    height: '440px',
+    ':hover': {
+      transform: 'scale(1.02)',
+    },
+    transition: '0.3s',
+    overflow: 'hidden',
+  },
+  like: {
+    display: 'flex',
+    gap: '5px',
+    justifyContent: 'flex-end',
+  },
+  header: {
+    display: 'flex',
+  },
+  img: {
+    width: '40px',
+    height: '40px',
+    objectFit: 'cover',
+    borderRadius: '50%',
+  },
+  name: {
+    lineHeight: '40px',
+    marginLeft: '10px',
+    fontSize: '13px',
+  },
+  title: {
+    fontSize: '22px',
+    fontWeight: '600',
+    margin: '20px 0',
+  },
+  content: {
+    height: '300px',
+    lineHeight: '27px',
+    overflow: 'hidden',
+  },
+});
+
+const Feed = ({ likeCount, link, title, content, name, profile }) => {
+  return (
+    <Link to={link}>
+      <article {...stylex.props(feed.box)}>
+        <div {...stylex.props(feed.like)}>
+          <span>
+            <i className="fa-solid fa-heart"></i>
+          </span>
+          <span>{likeCount}</span>
+        </div>
+        <div {...stylex.props(feed.header)}>
+          <img {...stylex.props(feed.img)} src={profile}></img>
+          <div {...stylex.props(feed.name)}>{name}</div>
+        </div>
+
+        <div {...stylex.props(feed.title)}>{title}</div>
+        <div {...stylex.props(feed.content)}>
+          {content && content.length > 350
+            ? `${content.slice(0, 350)}...`
+            : content}
+        </div>
+      </article>
+    </Link>
+  );
+};
+
+export default Feed;
